+++ data-service-base/src/main/java/org/fugerit/java/dsb/attributes/ServiceMap.java
@@ -1,0 +1,11 @@
+package org.fugerit.java.dsb.attributes;
+
+public interface ServiceMap<K,V> {
+
+    V get(K key);
+
+    void set(K key, V value);
+
+    void remove(K key);
+
+}