+++ data-service-base/src/main/java/org/fugerit/java/dsb/attributes/SimpleServiceMap.java
@@ -1,0 +1,5 @@
+package org.fugerit.java.dsb.attributes;
+
+public interface SimpleServiceMap extends ServiceMap<String, String> {
+
+}