+++ data-service-base/src/main/java/org/fugerit/java/dsb/attributes/impl/SimpleServiceMapDefault.java
@@ -1,0 +1,5 @@
+package org.fugerit.java.dsb.attributes.impl;
+
+public class SimpleServiceMapDefault extends ServiceMapDefault<String, String> implements org.fugerit.java.dsb.attributes.SimpleServiceMap {
+
+}